FastAPI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-groq==0.1.9
langchain-core==0.1.10
python-multipart==0.0.6
python-dotenv==1.0.0